Renamed and created folders also Fixed wrong space name
