+++ src/Teams.Business/Models/TeamBusiness.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+
 namespace Teams.Business.Models
 {
     public class TeamBusiness
